+++ packages/atomic/cypress/e2e/refine-toggle-actions.ts
@@ -3,6 +3,7 @@ import {
   TagProps,
   TestFixture,
 } from '../fixtures/test-fixture';
+import {automaticFacetGeneratorComponent} from './facets/automatic-facet-generator/automatic-facet-generator-assertions';
 import {hierarchicalField} from './facets/category-facet/category-facet-actions';
 import {categoryFacetComponent} from './facets/category-facet/category-facet-selectors';
 import {colorFacetField} from './facets/color-facet/color-facet-actions';
@@ -52,6 +53,26 @@ export const addRefineToggle =
     manager.append(timeframeFacet);
 
     env.withElement(manager).withElement(refineToggle);
+  };
+
+export const addRefineToggleWithoutFacets =
+  (props: TagProps = {}) =>
+  (env: TestFixture) => {
+    const refineToggle = generateComponentHTML(refineToggleComponent, props);
+
+    env.withElement(refineToggle);
+  };
+
+export const addRefineToggleWithAutomaticFacets =
+  (props: TagProps = {}) =>
+  (env: TestFixture) => {
+    const automaticFacetGenerator = generateComponentHTML(
+      automaticFacetGeneratorComponent,
+      {'desired-count': '3'}
+    );
+    const refineToggle = generateComponentHTML(refineToggleComponent, props);
+
+    env.withElement(automaticFacetGenerator).withElement(refineToggle);
   };
 
 export const addRefineToggleRangeVariations =